+++ simple-mysqli.php
@@ -19,6 +19,7 @@ class SimpleMySQLi {
 
 	/**
 	 * SimpleMySQLi constructor
+	 *
 	 * @param string $host Hostname or an IP address, like localhost or 127.0.0.1
 	 * @param string $username Database username
 	 * @param string $password Database password
@@ -32,7 +33,7 @@ class SimpleMySQLi {
 	 * @throws SimpleMySQLiException If $defaultFetchType specified isn't one of the allowed fetch modes
 	 * @throws mysqli_sql_exception If any mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
-	public function __construct(string $host, string $username, string $password, string $dbName, string $charset = 'utf8', string $defaultFetchType = 'assoc') {
+	public function __construct(string $host, string $username, string $password, string $dbName, string $charset = 'utf8mb4', string $defaultFetchType = 'assoc') {
 		$this->defaultFetchType = $defaultFetchType;
 
 		if(!in_array($defaultFetchType, self::ALLOWED_FETCH_TYPES_BOTH)) { //check if it is an allowed fetch type
@@ -47,13 +48,15 @@ class SimpleMySQLi {
 
 	/**
 	 * All queries go here. If select statement, needs to be used with either `fetch()` for single row and loop fetching or
-	 *`fetchAll()` for fetching all results.
+	 *`fetchAll()` for fetching all results
+	 *
 	 * @param string $sql SQL query
 	 * @param array $values (optional) Values or variables to bind to query. Can be empty for selecting all rows
 	 * @param string $types (optional) Variable type for each bound value/variable
+	 * @return $this
 	 * @throws mysqli_sql_exception If any mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
-	public function query(string $sql, array $values = [], string $types = '') {
+	public function query(string $sql, array $values = [], string $types = '') {		
 		if(!$types) $types = str_repeat('s', count($values)); //String type for all variables if not specified
 
 		$stmt = $this->stmt = $this->mysqli->prepare($sql);
@@ -63,9 +66,49 @@ class SimpleMySQLi {
 
 		return $this;
 	}
+	
+	/**
+	 * Used in order to be more efficient if same SQL is used with different values
+	 *
+	 * @param array $values (optional) Values or variables to bind to query. Can be empty for selecting all rows
+	 * @param string $types (optional) Variable type for each bound value/variable
+	 * @return $this
+	 * @throws mysqli_sql_exception If any mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function execute(array $values = [], string $types = '') {		
+		if(!$types) $types = str_repeat('s', count($values)); //String type for all variables if not specified
+		
+		$stmt = $this->stmt;
+		$stmt->bind_param($types, ...$values);
+		$stmt->execute();
+		
+		return $this;
+	}
+	
+	/**
+	 * Create correct number of questions marks for WHERE IN() array
+	 *
+	 * @param array $inArr Array used in WHERE IN clause
+	 * @return string Correct number of question marks
+	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function whereIn(array $inArr) {
+		return $clause = implode(',', array_fill(0, count($inArr), '?')); //create question marks
+	}
+	
+	/**
+	 * Get number of rows from SELECT
+	 *
+	 * @return int $mysqli->num_rows
+	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function numRows() {
+		return $this->stmtResult->num_rows;
+	}
 
 	/**
-	 * Get affected rows
+	 * Get affected rows. Can be used instead of numRows() in SELECT
+	 *
 	 * @return int $mysqli->affected_rows
 	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
@@ -76,6 +119,7 @@ class SimpleMySQLi {
 	/**
 	 * A more specific version of affectedRows() to give you more info what happened. Uses $mysqli::info under the hood
 	 * Can be used for the following cases http://php.net/manual/en/mysqli.info.php
+	 *
 	 * @return array Associative array converted from result string
 	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
@@ -86,6 +130,7 @@ class SimpleMySQLi {
 
 	/**
 	 * Get the latest primary key inserted
+	 *
 	 * @return int $mysqli->insert_id
 	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
@@ -95,13 +140,14 @@ class SimpleMySQLi {
 
 	/**
 	* Fetch one row at a time
+	*
 	* @param string $fetchType (optional) This overrides the default fetch type set in the constructor. Can be an default type
 	* @return mixed Array of either fetch type specified or default fetch mode. Can be a scalar too. Null if no more rows
 	* @throws SimpleMySQLiException If $fetchType specified isn't one of the allowed fetch modes in $defaultFetchType
 	*                               If fetch mode specification is violated
 	* @throws mysqli_sql_exception If any mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	*/
-	public function fetch(string $fetchType = '') {
+	public function fetch(string $fetchType = '', $className = '') {
 		$stmtResult = $this->stmtResult;
 		$row = [];
 
@@ -111,10 +157,15 @@ class SimpleMySQLi {
 			$allowedComma = implode("','", self::ALLOWED_FETCH_TYPES_BOTH);
 			throw new SimpleMySQLiException("The variable 'fetchType' must be '$allowedComma'. You entered '$fetchType'");
 		}
+		
+		if($fetchType !== 'obj' && $className) {
+			throw new SimpleMySQLiException("You can only specify a class name with 'obj' as the fetch type");
+		}
 
 		if($fetchType === 'num') $row = $stmtResult->fetch_row();
 		else if($fetchType === 'assoc') $row = $stmtResult->fetch_assoc();
-		else if($fetchType === 'obj') $row = $stmtResult->fetch_object();
+		else if($fetchType === 'obj' && !$className) $row = $stmtResult->fetch_object();
+		else if($fetchType === 'obj' && $className) $row = $stmtResult->fetch_object($className);
 		else if($fetchType === 'col') {
 			if($stmtResult->field_count !== 1) {
 				throw new SimpleMySQLiException("The fetch type: '$fetchType' must have exactly 1 column in query");
@@ -127,6 +178,7 @@ class SimpleMySQLi {
 
 	/**
 	* Fetch all results in array
+	*
 	* @param string $fetchType (optional) This overrides the default fetch type set in the constructor. Can be any default type and:
 	*               'keyPair' - Unique key (1st column) to single value (2nd column). Same as PDO::FETCH_KEY_PAIR
 	*               'keyPairArr' - Unique key (1st column) to array. Same as PDO::FETCH_UNIQUE
@@ -137,7 +189,7 @@ class SimpleMySQLi {
 	*                               If fetch mode specification is violated
 	* @throws mysqli_sql_exception If any mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	*/
-	public function fetchAll(string $fetchType = '') {
+	public function fetchAll(string $fetchType = '', $className = '') {
 		$stmtResult = $this->stmtResult;
 		$arr = [];
 
@@ -149,13 +201,24 @@ class SimpleMySQLi {
 			$allowedComma = implode("','", $comboAllowedTypes);
 			throw new SimpleMySQLiException("The variable 'fetchType' must be '$allowedComma'. You entered '$fetchType'");
 		}
+		
+		if($fetchType !== 'obj' && $className) {
+			throw new SimpleMySQLiException("You can only specify a class name with 'obj' as the fetch type");
+		}
 
 		//All of the fetch types
 		if($fetchType === 'num') $arr = $stmtResult->fetch_all(MYSQLI_NUM);
 		else if($fetchType === 'assoc') $arr = $stmtResult->fetch_all(MYSQLI_ASSOC);
 		else if($fetchType === 'obj') {
-			while($row = $stmtResult->fetch_object()) {
-				$arr[] = $row;
+			if(!$className) {
+				while($row = $stmtResult->fetch_object()) {
+					$arr[] = $row;
+				}
+			}
+			else {
+				while($row = $stmtResult->fetch_object($className)) {
+					$arr[] = $row;
+				}
 			}
 		}
 		else if($fetchType === 'col') {
@@ -185,8 +248,6 @@ class SimpleMySQLi {
 			}
 		}
 
-		$stmtResult->free();
-
 		return $arr;
 	}
 
@@ -207,7 +268,7 @@ class SimpleMySQLi {
 			if($types) $countTypes = count($types);
 
 			if(!is_array($sql)) {
-				$currSql = $sql;
+				$currSQL = $sql;
 				$isArray = false;
 			}
 			else $countSql = count($sql); //Only count sql if array
@@ -223,11 +284,16 @@ class SimpleMySQLi {
 				if(!$types) $currTypes = str_repeat('s', count($values[$x])); //String type for all variables if not specified
 				else $currTypes = $types[$x];
 
-				if($isArray) $currSql = $sql[$x]; //Either different queries or the same one with different values
-
-				$stmt = $this->mysqli->prepare($currSql);
+				if($isArray) $currSQL = $sql[$x]; //Either different queries or the same one with different values
+				
+				if($isArray || (!$isArray && $x === 0)) { //Make it more efficient if same query used multiple times with different values
+					$stmt = $this->mysqli->prepare($currSQL);
+				}
 				$stmt->bind_param($currTypes, ...$values[$x]);
 				$stmt->execute();
+				if($this->affectedRows() < 1) { 
+					throw new SimpleMySQLiException("Query did not succeed, with affectedRows() of: {$this->affectedRows()} Query: $currSQL");
+				}
 				$stmt->close();
 			}
 			$this->mysqli->autocommit(TRUE);
@@ -236,9 +302,46 @@ class SimpleMySQLi {
 			throw $e;
 		}
 	}
-
+	
+	/**
+	 * Do stuff inside of transaction
+	 *
+	 * @param callable $callback Closure to do transaction operations inside. Parameter value is $this
+	 * @throws mysqli_sql_exception If transaction failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function transactionCallback(callable $callback) {
+		try {
+			$this->mysqli->autocommit(FALSE);	
+			$callback($this);
+		} catch(Exception $e) {
+			$this->mysqli->rollback();
+			throw $e;
+		} finally {
+			$this->mysqli->autocommit(TRUE);
+		}
+	}
+	
+	/**
+	 * Frees MySQL result
+	 *
+	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function freeResult() {
+		$this->stmtResult->free();
+	}
+	
+	/**
+	 * Closes MySQL prepared statement
+	 *
+	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
+	 */
+	public function closeStmt() {
+		$this->stmt->close();
+	}
+	
 	/**
 	 * Closes MySQL connection
+	 *
 	 * @throws mysqli_sql_exception If mysqli function failed due to mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT)
 	 */
 	public function close() {